requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-multipart>=0.0.12

# Async Support
httpx>=0.27.0
aiohttp>=3.10.0

# LLM & Agent Frameworks - Anthropic
anthropic>=0.39.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0

# Vector Database & RAG
chromadb>=0.5.0
sentence-transformers>=3.0.0
tiktoken>=0.7.0

# Data Processing
pandas>=2.2.0
numpy>=2.0.0
beautifulsoup4>=4.12.0
pypdf2>=3.0.0
python-docx>=1.1.0
markdown>=3.6

# Database
redis>=5.0.0
motor>=3.5.0
pymongo>=4.8.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-dotenv>=1.0.0

# Observability
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
prometheus-client>=0.21.0
structlog>=24.4.0

# External Integrations
slack-sdk>=3.30.0
jira>=3.8.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
faker>=30.0.0

# Development
black>=24.8.0
isort>=5.13.0
mypy>=1.11.0
ruff>=0.6.0
pre-commit>=3.8.0